+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -7,8 +7,8 @@ import kitchenpos.menu.domain.Product;
 import kitchenpos.common.vo.Name;
 import kitchenpos.common.vo.Price;
 import kitchenpos.common.vo.Quantity;
-import kitchenpos.dto.MenuCreateRequest;
-import kitchenpos.dto.MenuResponse;
+import kitchenpos.menu.application.request.MenuCreateRequest;
+import kitchenpos.menu.application.response.MenuResponse;
 import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.menu.domain.ProductRepository;
