+++ src/main/java/kitchenpos/menu/application/ProductService.java
@@ -3,8 +3,8 @@ package kitchenpos.menu.application;
 import kitchenpos.menu.domain.Product;
 import kitchenpos.common.vo.Name;
 import kitchenpos.common.vo.Price;
-import kitchenpos.dto.ProductCreateRequest;
-import kitchenpos.dto.ProductResponse;
+import kitchenpos.menu.application.request.ProductCreateRequest;
+import kitchenpos.menu.application.response.ProductResponse;
 import kitchenpos.menu.domain.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
