+++ src/main/java/kitchenpos/order/application/request/OrderCreateOrderLineItemRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.order.application.request;
 
 public class OrderCreateOrderLineItemRequest {
