+++ src/main/java/kitchenpos/table/application/request/OrderTableCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.table.application.request;
 
 public class OrderTableCreateRequest {
