+++ src/main/java/kitchenpos/table/application/request/OrderTableEmptyUpdateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.table.application.request;
 
 public class OrderTableEmptyUpdateRequest {
