+++ src/main/java/kitchenpos/table/application/request/TableGroupCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.table.application.request;
 
 import java.util.List;
