+++ src/test/java/kitchenpos/menu/application/MenuQueryServiceTest.java
@@ -8,7 +8,7 @@ import kitchenpos.menu.domain.Product;
 import kitchenpos.common.vo.Name;
 import kitchenpos.common.vo.Price;
 import kitchenpos.common.vo.Quantity;
-import kitchenpos.dto.MenuResponse;
+import kitchenpos.menu.application.response.MenuResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
