+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -6,9 +6,9 @@ import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.Product;
 import kitchenpos.common.vo.Name;
 import kitchenpos.common.vo.Price;
-import kitchenpos.dto.MenuCreateRequest;
-import kitchenpos.dto.MenuProductCreateRequest;
-import kitchenpos.dto.MenuResponse;
+import kitchenpos.menu.application.request.MenuCreateRequest;
+import kitchenpos.menu.application.request.MenuProductCreateRequest;
+import kitchenpos.menu.application.response.MenuResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
